pyproject.toml
[project]
name = "boardflow"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=13.0.0",
    "alembic>=1.18.4",
    "anthropic>=0.25.0",
    "asyncpg>=0.29.0",
    "boto3>=1.35.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "pandas>=3.0.1",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.7",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.47",
    "tenacity>=9.1.4",
    "uvicorn>=0.32.0",
]
